refactor(acpx): split Windows command parsing
