+++ extensions/acpx/src/runtime-internals/mcp-proxy.mjs
@@ -4,116 +4,7 @@ import { spawn } from "node:child_process";
 import path from "node:path";
 import { createInterface } from "node:readline";
 import { pathToFileURL } from "node:url";
-
-const WINDOWS_EXECUTABLE_PATH_RE =
-  /^(?<command>(?:[A-Za-z]:[\\/]|\\\\[^\\/]+[\\/][^\\/]+[\\/]).*?\.(?:exe|com))(?=\s|$)(?:\s+(?<rest>.*))?$/i;
-
-function splitCommandParts(value, platform = process.platform) {
-  const parts = [];
-  let current = "";
-  let quote = null;
-  let escaping = false;
-
-  for (let index = 0; index < value.length; index += 1) {
-    const ch = value[index];
-    const next = value[index + 1];
-    if (escaping) {
-      current += ch;
-      escaping = false;
-      continue;
-    }
-    if (ch === "\\") {
-      if (quote === "'") {
-        current += ch;
-        continue;
-      }
-      if (platform === "win32") {
-        if (quote === '"') {
-          if (next === '"' || next === "\\") {
-            escaping = true;
-            continue;
-          }
-          current += ch;
-          continue;
-        }
-        if (!quote) {
-          current += ch;
-          continue;
-        }
-      }
-      escaping = true;
-      continue;
-    }
-    if (quote) {
-      if (ch === quote) {
-        quote = null;
-      } else {
-        current += ch;
-      }
-      continue;
-    }
-    if (ch === "'" || ch === '"') {
-      quote = ch;
-      continue;
-    }
-    if (/\s/.test(ch)) {
-      if (current.length > 0) {
-        parts.push(current);
-        current = "";
-      }
-      continue;
-    }
-    current += ch;
-  }
-
-  if (escaping) {
-    current += "\\";
-  }
-  if (quote) {
-    throw new Error("Invalid agent command: unterminated quote");
-  }
-  if (current.length > 0) {
-    parts.push(current);
-  }
-  if (parts.length === 0) {
-    return [];
-  }
-  return parts;
-}
-
-function splitWindowsExecutableCommand(value, platform = process.platform) {
-  if (platform !== "win32") {
-    return null;
-  }
-  const trimmed = value.trim();
-  if (!trimmed || trimmed.startsWith('"') || trimmed.startsWith("'")) {
-    return null;
-  }
-  const match = trimmed.match(WINDOWS_EXECUTABLE_PATH_RE);
-  if (!match?.groups?.command) {
-    return null;
-  }
-  const rest = match.groups.rest?.trim() ?? "";
-  return {
-    command: match.groups.command,
-    args: rest ? splitCommandParts(rest, platform) : [],
-  };
-}
-
-export function splitCommandLine(value, platform = process.platform) {
-  const windowsCommand = splitWindowsExecutableCommand(value, platform);
-  if (windowsCommand) {
-    return windowsCommand;
-  }
-  const parts = splitCommandParts(value, platform);
-  if (parts.length === 0) {
-    throw new Error("Invalid agent command: empty command");
-  }
-  return {
-    command: parts[0],
-    args: parts.slice(1),
-  };
-}
+import { splitCommandLine } from "./mcp-command-line.mjs";
 
 function decodePayload(argv) {
   const payloadIndex = argv.indexOf("--payload");
